Move read_password to add.rs
- this is the only place it is being used
- also, this help with a cyclic dependency i was noticing when trying to
use secure_store in secret and generate
