+++ cmd/soroban-cli/src/config/secret.rs
@@ -1,20 +1,18 @@
 use clap::arg;
 use serde::{Deserialize, Serialize};
-use std::{io::Write, str::FromStr};
+use std::str::FromStr;
 
 use sep5::SeedPhrase;
 use stellar_strkey::ed25519::{PrivateKey, PublicKey};
 
 use crate::{
-    config::address::KeyName, print::Print, signer::{self, keyring, secure_store::SecureStore, LocalKey, SecureStoreEntry, Signer, SignerKind}, utils
+    print::Print, signer::{self, keyring, LocalKey, SecureStoreEntry, Signer, SignerKind}, utils
 };
 
 #[derive(thiserror::Error, Debug)]
 pub enum Error {
     // #[error("seed_phrase must be 12 words long, found {len}")]
     // InvalidSeedPhrase { len: usize },
-    #[error("secret input error")]
-    PasswordRead,
     #[error(transparent)]
     Secret(#[from] stellar_strkey::DecodeError),
     #[error(transparent)]
@@ -29,8 +27,6 @@ pub enum Error {
     Keyring(#[from] keyring::Error),
     #[error("Secure Store does not reveal secret key")]
     SecureStoreDoesNotRevealSecretKey,
-    #[error("Getting key from secure store failed")] // Todo: update this
-    SecureStore,
 }
 
 #[derive(Debug, clap::Args, Clone)]
@@ -45,26 +41,6 @@ pub struct Args {
     /// Save the new key in secure store. This only supports seed phrases for now.
     #[arg(long)]
     pub secure_store: bool,
-}
-
-impl Args {
-    pub fn read_secret(&self, name: &KeyName) -> Result<Secret, Error> {
-        if let Ok(secret_key) = std::env::var("SOROBAN_SECRET_KEY") {
-            return Ok(Secret::SecretKey { secret_key })
-        }
-
-        println!("Type a secret key or 12/24 word seed phrase:");
-        let secret_key = read_password()?;
-        if self.secure_store {
-            let seed_phrase: SeedPhrase = secret_key.parse()?;
-            let print = &Print::new(false);
-            SecureStore::secret(print, name, seed_phrase).map_err(|_| Error::SecureStore)
-        } else {
-            secret_key
-                .parse()
-                .map_err(|_| Error::InvalidSecretOrSeedPhrase)
-        }
-    }
 }
 
 #[derive(Debug, Serialize, Deserialize)]
@@ -178,11 +154,6 @@ pub fn seed_phrase_from_seed(seed: Option<&str>) -> Result<SeedPhrase, Error> {
 
 pub fn test_seed_phrase() -> Result<SeedPhrase, Error> {
     Ok("0000000000000000".parse()?)
-}
-
-fn read_password() -> Result<String, Error> {
-    std::io::stdout().flush().map_err(|_| Error::PasswordRead)?;
-    rpassword::read_password().map_err(|_| Error::PasswordRead)
 }
 
 #[cfg(test)]
